+++ CHANGELOG.md
@@ -9,6 +9,7 @@ Go to the `v1` branch to see the changelog of Lume 1.
 ### Added
 - New plugin `sri`.
 - Improved plugin docs with links to the online documentation.
+- New functions `nav.nextPage()` and `nav.previousPage()` for `nav` plugin.
 
 ### Fixed
 - Updated deps: `std`, `liquidjs`, `preact`, `tailwindcss`, `xml`, `postcss`, `autoprefixer`, `unocss`.